backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
google-genai==1.14.0
pillow==11.1.0
python-multipart==0.0.20
pydantic-settings==2.7.1
pydantic==2.10.5
aiofiles==24.1.0
